FE: Data logic moved to containers, some refactoring, comments.
